+++ skrub/_table_vectorizer.py
@@ -1,5 +1,4 @@
 import reprlib
-import warnings
 from collections import UserDict
 from typing import Iterable
 
@@ -121,6 +120,7 @@ def _get_preprocessors(
     drop_if_constant,
     n_jobs,
     add_tofloat32=True,
+    datetime_format=None,
 ):
     steps = [CheckInputDataFrame()]
     transformers = [
@@ -130,7 +130,7 @@ def _get_preprocessors(
             drop_if_constant=drop_if_constant,
             drop_if_unique=drop_if_unique,
         ),
-        ToDatetime(),
+        ToDatetime(format=datetime_format),
     ]
     if add_tofloat32:
         transformers.append(ToFloat32())
@@ -179,6 +179,9 @@ class Cleaner(TransformerMixin, BaseEstimator):
         of unique values is equal to the number of rows in the column. Numeric columns
         are never dropped.
 
+    datetime_format : str, default=None
+        The format to use when parsing dates. If None, the format is inferred.
+
     n_jobs : int, default=None
         Number of jobs to run in parallel.
         ``None`` means 1 unless in a joblib ``parallel_backend`` context.
@@ -203,12 +206,16 @@ class Cleaner(TransformerMixin, BaseEstimator):
     - ``CleanNullStrings()``: replace strings used to represent null values
       with actual null values.
 
-    - ``DropUninformative()``: drop the column if it contains too many null values,
-    if it contains only one distinct value, or if all values are distinct (off by
-    default).
+    - ``DropUninformative()``: drop the column if it is considered to be
+      "uninformative". A column is considered to be "uninformative" if it contains
+      only null values (``drop_null_fraction``), only a constant value
+      (``drop_if_constant``), or if all values are distinct (``drop_if_unique``).
+      Note that setting ``drop_if_unique`` to ``True`` may lead to dropping columns
+      that contain text.
 
     - ``ToDatetime()``: parse datetimes represented as strings and return them as
-      actual datetimes with the correct dtype.
+      actual datetimes with the correct dtype. If ``datetime_format`` is provided,
+      it is forwarded to ``ToDatetime()``. Otherwise, the format is inferred.
 
     - ``CleanCategories()``: process categorical columns depending on the dataframe
       library (Pandas or Polars) to force consistent typing and avoid issues downstream.
@@ -286,11 +293,13 @@ class Cleaner(TransformerMixin, BaseEstimator):
         drop_null_fraction=1.0,
         drop_if_constant=False,
         drop_if_unique=False,
+        datetime_format=None,
         n_jobs=1,
     ):
         self.drop_null_fraction = drop_null_fraction
         self.drop_if_constant = drop_if_constant
         self.drop_if_unique = drop_if_unique
+        self.datetime_format = datetime_format
         self.n_jobs = n_jobs
 
     def fit_transform(self, X, y=None):
@@ -318,6 +327,7 @@ class Cleaner(TransformerMixin, BaseEstimator):
             drop_if_unique=self.drop_if_unique,
             n_jobs=self.n_jobs,
             add_tofloat32=False,
+            datetime_format=self.datetime_format,
         )
         self._pipeline = make_pipeline(*all_steps)
         result = self._pipeline.fit_transform(X)
@@ -363,18 +373,6 @@ class Cleaner(TransformerMixin, BaseEstimator):
         """
         self.fit_transform(X, y=y)
         return self
-
-
-class SimpleCleaner(Cleaner):
-    def __init__(self, drop_null_fraction=1.0, n_jobs=1):
-        super().__init__(drop_null_fraction=drop_null_fraction, n_jobs=n_jobs)
-        warnings.warn(
-            (
-                "SimpleCleaner was renamed to Cleaner and will be removed in the "
-                "next release."
-            ),
-            category=DeprecationWarning,
-        )
 
 
 class TableVectorizer(TransformerMixin, BaseEstimator):
@@ -474,6 +472,9 @@ class TableVectorizer(TransformerMixin, BaseEstimator):
         If set to true, drop columns that contain only unique values, i.e., the number
         of unique values is equal to the number of rows in the column. Numeric columns
         are never dropped.
+
+    datetime_format : str, default=None
+        The format to use when parsing dates. If None, the format is inferred.
 
     n_jobs : int, default=None
         Number of jobs to run in parallel.
@@ -600,6 +601,9 @@ class TableVectorizer(TransformerMixin, BaseEstimator):
     represented as strings. By default, columns that contain only null values are
     dropped. Moreover, a final post-processing step is applied to all
     non-categorical columns in the encoder's output to cast them to float32.
+    If ``datetime_format`` is provided, it will be used to parse all datetime
+    columns.
+
     We can inspect all the processing steps that were applied to a given column:
 
     >>> vectorizer.all_processing_steps_['B']
@@ -704,6 +708,7 @@ class TableVectorizer(TransformerMixin, BaseEstimator):
         drop_null_fraction=1.0,
         drop_if_constant=False,
         drop_if_unique=False,
+        datetime_format=None,
         n_jobs=None,
     ):
         self.cardinality_threshold = cardinality_threshold
@@ -720,6 +725,7 @@ class TableVectorizer(TransformerMixin, BaseEstimator):
         self.drop_null_fraction = drop_null_fraction
         self.drop_if_constant = drop_if_constant
         self.drop_if_unique = drop_if_unique
+        self.datetime_format = datetime_format
 
     def fit(self, X, y=None):
         """Fit transformer.
@@ -840,7 +846,7 @@ class TableVectorizer(TransformerMixin, BaseEstimator):
         )
 
         transformer_list += [
-            ToDatetime(),
+            ToDatetime(format=self.datetime_format),
             ToFloat32(),
             CleanCategories(),
             ToStr(),