+++ CHANGES.rst
@@ -56,6 +56,12 @@ Changes
 - The :func:`concat_horizontal` function was replaced with :func:`concat`. Horizontal or vertical concatenation
   is now controlled by the `axis` parameter. :pr:`1334` by :user:`Parasa V Prajwal <pvprajwal>`.
 
+- The :class:`TableVectorizer` and :class:`Cleaner` now accept a `datetime_format`
+  parameter for specifying the format to use when parsing datetime columns.
+  :pr:`1358` by :user:`Riccardo Cappuzzo<rcap107>`.
+
+- The :class:`SimpleCleaner` has been removed. use :class:`Cleaner` instead. :pr:`1370` by :user:`Riccardo Cappuzzo<rcap107>`.
+
 Bugfixes
 --------